requirements.txt
cffi
gunicorn
psycopg2==2.6.1
flask==0.10.1
pyScss==1.3.4
Flask-Assets==0.11
Flask-SQLAlchemy==2.0
alembic==0.8.2
Flask-WTF==0.12
flask-login==0.3.2
flask-bcrypt==0.7.1
tinys3==0.1.11
Flask-Admin==1.4.0
requests==2.10.0
